gim/pyproject.toml
[project]
name = "gim"
version = "0.1.0"
description = "Global Issue Memory - Privacy-preserving MCP server for AI coding issue intelligence"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.0.0",
    "supabase>=2.0.0",
    "qdrant-client>=1.7.0",
    "google-genai>=1.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
